+++ src/main/kotlin/com/github/victorrentea/livecoding/extracthints/LocalUsageType.kt
@@ -1,0 +1,5 @@
+package com.github.victorrentea.livecoding.extracthints
+
+enum class LocalUsageType {
+    READ, WRITE
+}