+++ src/test/live-coding-playground/src/main/java/introduceguard/MissingTopLevelElse.java
@@ -5,9 +5,14 @@ public class MissingTopLevelElse {
         if (x != 0) { //
             for (int i=0;i<x;i++) {
                 if (i == x%2) {
-                    System.out.println("Stuff to do");
+                    System.out.println("Stuff to do".concat("a"));
                     if (x == 5) {
                         System.out.println("More stuff to do");
+
+                        System.out.println("Bla bla ");
+                        System.out.println("Bla bla ");
+                        System.out.println("Bla bla ");
+                        System.out.println("Bla bla ");
                     }
                 }
             }